# openapi.yaml
openapi: 3.0.3
info:
  title: Stock Watchlist API
  description: |
    A comprehensive REST API for managing stock watchlists with real-time price notifications in a terminal application.
    
    This API allows users to:
    - Create and manage multiple watchlists
    - Add and remove stocks from watchlists
    - Track stock information including ticker symbols and prices
    - Set price targets and receive FCM notifications
    - Monitor watchlist performance with real-time WebSocket price feeds
    
    The API uses Firebase Firestore for data persistence, WebSocket for real-time price updates, and FCM for push notifications.
  version: 1.0.0
  contact:
    name: Stock Watchlist API Support
    email: support@musaffa.us
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://terminal.musaffa.us
    description: Production server
  - url: http://localhost:3000
    description: Development server
paths:
...
  /watchlists:
    get:
      summary: Get All Watchlists
      description: Retrieve a list of all watchlists for the authenticated user
      tags:
        - Watchlists
      responses:
        '200':
          description: Successfully retrieved watchlists
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [success]
                    example: success
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Watchlist'
                    description: Array of user's watchlists
                  count:
                    type: integer
                    minimum: 0
                    example: 3
                    description: Total number of watchlists
                  default_watchlist_id:
                    type: string
                    nullable: true
                    example: "watchlist_1758125631041"
                    description: ID of the user's default watchlist (null if none set)
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create New Watchlist
      description: Create a new watchlist with the specified name
      tags:
        - Watchlists
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                  example: "My Tech Portfolio"
                  description: Name of the watchlist
            examples:
              tech_portfolio:
                summary: Technology Portfolio
                value:
                  name: "Technology Stocks"
              energy_sector:
                summary: Energy Sector
                value:
                  name: "Energy Sector Watchlist"
      responses:
        '201':
          description: Watchlist created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [success]
                    example: success
                  message:
                    type: string
                    example: "Watchlist created successfully"
                  data:
                    $ref: '#/components/schemas/Watchlist'
        '400':
          description: Bad request - Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /user/preferences:
    get:
      summary: Get User Preferences
      description: Retrieve user preferences including default watchlist settings
      tags:
        - User Preferences
      responses:
        '200':
          description: User preferences retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [success]
                    example: success
                  data:
                    $ref: '#/components/schemas/UserPreferences'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /fcm/register-token:
    post:
      summary: Register FCM Token
      description: Register a Firebase Cloud Messaging token for push notifications
      tags:
        - FCM Tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  example: "fcm_device_token_here"
                  description: FCM device token from client
                device_type:
                  type: string
                  enum: [android, ios, web]
                  example: "android"
                  description: Type of device
                device_name:
                  type: string
                  example: "My Phone"
                  description: Human-readable device name
            examples:
              android_device:
                summary: Android Device
                value:
                  token: "fcm_android_token_123"
                  device_type: "android"
                  device_name: "Samsung Galaxy S21"
              ios_device:
                summary: iOS Device
                value:
                  token: "fcm_ios_token_456"
                  device_type: "ios"
                  device_name: "iPhone 13 Pro"
      responses:
        '201':
          description: FCM token registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [success]
                    example: success
                  message:
                    type: string
                    example: "FCM token registered successfully"
                  data:
                    type: object
                    properties:
                      token_id:
                        type: string
                        example: "uuid-here"
                      device_type:
                        type: string
                        example: "android"
                      device_name:
                        type: string
                        example: "My Phone"
                      created_at:
                        type: string
                        format: date-time
                        example: "2024-01-17T10:30:00.000Z"
        '400':
          description: Bad request - Missing FCM token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /fcm/tokens:
    get:
      summary: Get User FCM Tokens
      description: Retrieve all registered FCM tokens for the user
      tags:
        - FCM Tokens
      responses:
        '200':
          description: FCM tokens retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [success]
                    example: success
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        token_id:
                          type: string
                          example: "uuid-here"
                        device_type:
                          type: string
                          example: "android"
                        device_name:
                          type: string
                          example: "My Phone"
                        created_at:
                          type: string
                          format: date-time
                          example: "2024-01-17T10:30:00.000Z"
                        last_used:
                          type: string
                          format: date-time
                          example: "2024-01-17T10:30:00.000Z"
                  count:
                    type: integer
                    example: 2
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /fcm/tokens/{tokenId}:
    put:
      summary: Update FCM Token
      description: Update FCM token information
      tags:
        - FCM Tokens
      parameters:
        - name: tokenId
          in: path
          required: true
          description: Unique identifier of the FCM token
          schema:
            type: string
          example: "uuid-here"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                device_name:
                  type: string
                  example: "Updated Device Name"
                is_active:
                  type: boolean
                  example: true
            examples:
              update_device_name:
                summary: Update Device Name
                value:
                  device_name: "My Updated Phone"
              deactivate_token:
                summary: Deactivate Token
                value:
                  is_active: false
      responses:
        '200':
          description: FCM token updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [success]
                    example: success
                  message:
                    type: string
                    example: "FCM token updated successfully"
                  data:
                    type: object
                    properties:
                      token_id:
                        type: string
                        example: "uuid-here"
                      last_updated:
                        type: string
                        format: date-time
                        example: "2024-01-17T10:30:00.000Z"
        '404':
          description: FCM token not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Remove FCM Token
      description: Remove (deactivate) an FCM token
      tags:
        - FCM Tokens
      parameters:
        - name: tokenId
          in: path
          required: true
          description: Unique identifier of the FCM token
          schema:
            type: string
          example: "uuid-here"
      responses:
        '200':
          description: FCM token removed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [success]
                    example: success
                  message:
                    type: string
                    example: "FCM token removed successfully"
                  data:
                    type: object
                    properties:
                      token_id:
                        type: string
                        example: "uuid-here"
        '404':
          description: FCM token not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /targets:
    post:
      summary: Create Price Target
      description: Create a new price target for stock price notifications
      tags:
        - Price Targets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - ticker
                - target_price
                - alert_type
                - watchlist_id
              properties:
                ticker:
                  type: string
                  pattern: '^[A-Z]{1,5}$'
                  example: "AAPL"
                  description: Stock ticker symbol
                target_price:
                  type: number
                  format: float
                  minimum: 0
                  example: 150.00
                  description: Target price to monitor
                alert_type:
                  type: string
                  enum: [above, below]
                  example: "above"
                  description: Alert when price goes above or below target
                watchlist_id:
                  type: string
                  example: "watchlist_1758125631041"
                  description: ID of the watchlist this target belongs to
                is_active:
                  type: boolean
                  example: true
                  description: Whether the target is active
            examples:
              above_target:
                summary: Price Above Target
                value:
                  ticker: "AAPL"
                  target_price: 150.00
                  alert_type: "above"
                  watchlist_id: "watchlist_1758125631041"
                  is_active: true
              below_target:
                summary: Price Below Target
                value:
                  ticker: "TSLA"
                  target_price: 200.00
                  alert_type: "below"
                  watchlist_id: "watchlist_1758125631042"
                  is_active: true
      responses:
        '201':
          description: Price target created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [success]
                    example: success
                  message:
                    type: string
                    example: "Price target created successfully"
                  data:
                    type: object
                    properties:
                      target_id:
                        type: string
                        example: "uuid-here"
                      ticker:
                        type: string
                        example: "AAPL"
                      target_price:
                        type: number
                        example: 150.00
                      alert_type:
                        type: string
                        example: "above"
                      is_active:
                        type: boolean
                        example: true
                      created_at:
                        type: string
                        format: date-time
                        example: "2024-01-17T10:30:00.000Z"
        '400':
          description: Bad request - Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict - Active target already exists for this ticker
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: Get All Price Targets
      description: Retrieve all active price targets for the user
      tags:
        - Price Targets
      responses:
        '200':
          description: Price targets retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [success]
                    example: success
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        target_id:
                          type: string
                          example: "uuid-here"
                        ticker:
                          type: string
                          example: "AAPL"
                        target_price:
                          type: number
                          example: 150.00
                        alert_type:
                          type: string
                          example: "above"
                        is_active:
                          type: boolean
                          example: true
                        created_at:
                          type: string
                          format: date-time
                          example: "2024-01-17T10:30:00.000Z"
                        last_updated:
                          type: string
                          format: date-time
                          example: "2024-01-17T10:30:00.000Z"
                        triggered:
                          type: boolean
                          example: false
                        triggered_at:
                          type: string
                          format: date-time
                          nullable: true
                          example: null
                        current_price:
                          type: number
                          example: 145.25
                  count:
                    type: integer
                    example: 3
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /notifications:
    get:
      summary: Get Notification History
      description: Retrieve notification history for the user
      tags:
        - Notifications
      parameters:
        - name: limit
          in: query
          description: Maximum number of notifications to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          example: 50
      responses:
        '200':
          description: Notification history retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [success]
                    example: success
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: "notification-id"
                        target_id:
                          type: string
                          example: "target-uuid"
                        ticker:
                          type: string
                          example: "AAPL"
                        target_price:
                          type: number
                          example: 150.00
                        current_price:
                          type: number
                          example: 150.25
                        alert_type:
                          type: string
                          example: "above"
                        sent_at:
                          type: string
                          format: date-time
                          example: "2024-01-17T10:30:00.000Z"
                        status:
                          type: string
                          example: "sent"
                  count:
                    type: integer
                    example: 5
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Watchlist:
      type: object
      description: A collection of stocks that a user wants to monitor
      properties:
        id:
          type: string
          example: "watchlist_1758125631041"
          description: Unique identifier for the watchlist
        name:
          type: string
          example: "My Tech Portfolio"
          description: Human-readable name of the watchlist
        date_created:
          type: string
          format: date-time
          example: "2024-01-17T10:30:00.000Z"
          description: Timestamp when the watchlist was created
        stock_count:
          type: integer
          minimum: 0
          example: 5
          description: Number of stocks currently in the watchlist
        is_default:
          type: boolean
          example: true
          description: Whether this watchlist is set as the user's default
      required:
        - id
        - name
        - date_created
        - stock_count
        - is_default
...
    UserPreferences:
      type: object
      description: User preferences including default watchlist settings
      properties:
        user_id:
          type: string
          example: "user123"
          description: Unique identifier for the user
        default_watchlist_id:
          type: string
          nullable: true
          example: "watchlist_1758125631041"
          description: ID of the user's default watchlist (null if none set)
        date_set:
          type: string
          format: date-time
          nullable: true
          example: "2024-01-17T10:30:00.000Z"
          description: Timestamp when the default watchlist was first set
        last_updated:
          type: string
          format: date-time
          nullable: true
          example: "2024-01-17T10:30:00.000Z"
          description: Timestamp when preferences were last updated
      required:
        - user_id
        - default_watchlist_id
        - date_set
        - last_updated
    Error:
      type: object
      description: Error response format
      properties:
        status:
          type: string
          enum: [error]
          example: error
          description: Indicates an error occurred
        message:
          type: string
          example: "Watchlist not found"
          description: Human-readable error message
      required:
        - status
        - message
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for API authentication
security:
  - bearerAuth: []
...